+++ src/main/java/pl/edu/wat/crochetshopapi/service/OrderService.java
@@ -34,7 +34,7 @@ public class OrderService {
 
         //Cart value
         List<Long> ids = new ArrayList<>();
-        long value = 0;
+        double value = 0;
         for (Product product : client.getCart().getProducts()) {
             value = value + product.getPrice();
             ids.add(product.getId());
